+++ flutter/example/integration_test/jason.dart
@@ -241,7 +241,7 @@ void main() {
     expect(
         () => conn.getRemoteMemberId(),
         throwsA(allOf(
-            // isStateError,
+            isStateError,
             predicate((e) =>
                 e is StateError &&
                 e.message == 'ConnectionHandle is in detached state.'))));
@@ -334,7 +334,7 @@ void main() {
             isStateError,
             predicate((e) =>
                 e is StateError &&
-                e.message == 'RoomHandle is in detached state.')));
+                e.message == 'RoomHandle is in detached state')));
 
     var formatExc;
     try {