+++ tests/room/room.rs
@@ -17,17 +17,27 @@ use medea_client_api_proto::{
     VideoSettings,
 };
 use medea_jason::{
-    api, media::MediaKind, peer::PeerConnection, room::Room,
-    rpc::MockRpcSession, utils::Updatable,
+    api::{
+        self,
+        err::{
+            LocalMediaInitException, LocalMediaInitExceptionKind, StateError,
+        },
+    },
+    media::MediaKind,
+    peer::PeerConnection,
+    room::Room,
+    rpc::MockRpcSession,
+    utils::Updatable,
 };
+use wasm_bindgen::JsValue;
 use wasm_bindgen_futures::{spawn_local, JsFuture};
 use wasm_bindgen_test::*;
 
 use crate::{
-    delay_for, get_constraints_update_exception, get_jason_error,
-    get_test_recv_tracks, get_test_required_tracks, get_test_tracks,
+    delay_for, get_test_recv_tracks, get_test_required_tracks, get_test_tracks,
     get_test_unrequired_tracks, media_stream_settings, timeout,
-    wait_and_check_test_result, yield_now, MockNavigator, TEST_ROOM_URL,
+    unchecked_jsval_cast, wait_and_check_test_result, yield_now, MockNavigator,
+    TEST_ROOM_URL,
 };
 
 wasm_bindgen_test_configure!(run_in_browser);
@@ -138,14 +148,18 @@ async fn error_get_local_stream_on_new_peer() {
     ))
     .await
     .unwrap();
+    let (cb, test_result) = js_callback!(|err: JsValue| {
+        let err: LocalMediaInitException = unchecked_jsval_cast(err);
+        let cause = err.cause().unwrap();
 
-    let (cb, test_result) = js_callback!(|err: api::Error| {
-        cb_assert_eq!(&err.name(), "UpdateLocalStreamError");
+        assert_eq!(err.kind(), LocalMediaInitExceptionKind::GetUserMediaFailed);
         cb_assert_eq!(
             &err.message(),
             "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Unknown JS error: error_get_local_stream_on_new_peer"
+             Error: error_get_local_stream_on_new_peer"
         );
+        assert_eq!(&cause.message(), "error_get_local_stream_on_new_peer");
+        assert!(&err.trace().contains("at src"));
     });
 
     room_handle.on_failed_local_media(cb.into()).unwrap();
@@ -185,13 +199,12 @@ async fn error_join_room_without_on_failed_stream_callback() {
         .on_connection_loss(js_sys::Function::new_no_args(""))
         .unwrap();
 
-    let err = get_jason_error(
+    let err: StateError = unchecked_jsval_cast(
         JsFuture::from(room_handle.join(String::from(TEST_ROOM_URL)))
             .await
             .unwrap_err(),
     );
 
-    assert_eq!(err.name(), "CallbackNotSet");
     assert_eq!(
         err.message(),
         "`Room.on_failed_local_media()` callback isn't set.",
@@ -215,13 +228,12 @@ async fn error_join_room_without_on_connection_loss_callback() {
         .on_failed_local_media(js_sys::Function::new_no_args(""))
         .unwrap();
 
-    let err = get_jason_error(
+    let err: StateError = unchecked_jsval_cast(
         JsFuture::from(room_handle.join(String::from(TEST_ROOM_URL)))
             .await
             .unwrap_err(),
     );
 
-    assert_eq!(err.name(), "CallbackNotSet");
     assert_eq!(
         err.message(),
         "`Room.on_connection_loss()` callback isn't set.",
@@ -1833,16 +1845,15 @@ async fn no_updates_sent_if_gum_fails_on_enable() {
 
     mock.error_get_user_media("gum error".into());
 
-    let err = JsFuture::from(room_handle.enable_audio())
-        .await
-        .unwrap_err();
-    let e = get_jason_error(err);
-    assert_eq!(e.name(), "CouldNotGetLocalMedia");
-    assert_eq!(
-        e.message(),
-        "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-         Unknown JS error: gum error",
+    let err: LocalMediaInitException = unchecked_jsval_cast(
+        JsFuture::from(room_handle.enable_audio())
+            .await
+            .unwrap_err(),
     );
+
+    assert_eq!(err.kind(), LocalMediaInitExceptionKind::GetUserMediaFailed);
+    assert!(err.message().contains("gum error"));
+
     mock.stop();
     drop(room);
 
@@ -1880,16 +1891,18 @@ async fn set_media_state_return_media_error() {
 
     mock.error_get_user_media(ERROR_MSG.into());
 
-    let err = get_jason_error(
+    let err: LocalMediaInitException = unchecked_jsval_cast(
         JsFuture::from(room_handle.enable_audio())
             .await
             .unwrap_err(),
     );
+
+    assert_eq!(err.kind(), LocalMediaInitExceptionKind::GetUserMediaFailed);
     assert_eq!(
         err.message(),
         format!(
             "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Unknown JS error: {}",
+             Error: {}",
             ERROR_MSG
         )
     );
@@ -2007,15 +2020,16 @@ async fn send_enabling_holds_local_tracks() {
 
     let mock = MockNavigator::new();
     mock.error_get_user_media("foobar".into());
-    let err = get_jason_error(
+    let err: LocalMediaInitException = unchecked_jsval_cast(
         JsFuture::from(room_handle.enable_video(None))
             .await
             .unwrap_err(),
     );
+    assert_eq!(err.kind(), LocalMediaInitExceptionKind::GetUserMediaFailed);
     assert_eq!(
         err.message(),
         "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-         Unknown JS error: foobar"
+         Error: foobar"
     );
     mock.stop();
 }
@@ -2023,6 +2037,11 @@ async fn send_enabling_holds_local_tracks() {
 /// Tests for [`RoomHandle::set_local_media_settings`].
 mod set_local_media_settings {
     use super::*;
+
+    use medea_jason::api::err::{
+        LocalMediaInitException, MediaSettingsUpdateException,
+        MediaStateTransitionException,
+    };
 
     /// Sets up connection between two peers in single room with first peer
     /// sending video to second peer.
@@ -2180,8 +2199,8 @@ mod set_local_media_settings {
         let (room, _rx) = get_test_room(Box::pin(event_rx));
         let room_handle = api::RoomHandle::from(room.new_handle());
 
-        let (cb, test_result) = js_callback!(|err: api::Error| {
-            cb_assert_eq!(&err.name(), "SenderCreateError");
+        let (cb, test_result) = js_callback!(|err: JsValue| {
+            let err: MediaStateTransitionException = unchecked_jsval_cast(err);
             cb_assert_eq!(
                 err.message(),
                 "MediaExchangeState of Sender cannot transit to \
@@ -2226,8 +2245,8 @@ mod set_local_media_settings {
     ///     1. `on_failed_local_media` was invoked.
     #[wasm_bindgen_test]
     async fn error_inject_invalid_local_stream_into_room_on_exists_peer() {
-        let (cb, test_result) = js_callback!(|err: api::Error| {
-            cb_assert_eq!(&err.name(), "UpdateLocalStreamError");
+        let (cb, test_result) = js_callback!(|err: JsValue| {
+            let err: MediaStateTransitionException = unchecked_jsval_cast(err);
             cb_assert_eq!(
                 &err.message(),
                 "provided multiple device video MediaStreamTracks"
@@ -2243,7 +2262,7 @@ mod set_local_media_settings {
 
         let room_handle = api::RoomHandle::from(room.new_handle());
         room_handle.on_failed_local_media(cb.into()).unwrap();
-        let err = get_constraints_update_exception(
+        let err: MediaSettingsUpdateException = unchecked_jsval_cast(
             JsFuture::from(room_handle.set_local_media_settings(
                 &constraints,
                 false,
@@ -2252,11 +2271,12 @@ mod set_local_media_settings {
             .await
             .unwrap_err(),
         );
-        let err = err.error().unwrap();
-        assert_eq!(err.name(), "InvalidLocalTracks");
+        let cause: MediaStateTransitionException =
+            unchecked_jsval_cast(err.cause().into());
+        assert_eq!(err.rolled_back(), false);
         assert_eq!(
-            err.message(),
-            "provided multiple device video MediaStreamTracks"
+            cause.message(),
+            "provided multiple device video MediaStreamTracks",
         );
 
         wait_and_check_test_result(test_result, || {}).await;
@@ -2395,7 +2415,7 @@ mod set_local_media_settings {
         let room_handle = api::RoomHandle::from(room.new_handle());
         let mock_navigator = MockNavigator::new();
         mock_navigator.error_get_user_media("disables_on_fail".into());
-        let err = get_constraints_update_exception(
+        let err: MediaSettingsUpdateException = unchecked_jsval_cast(
             JsFuture::from(room_handle.set_local_media_settings(
                 &media_settings_with_device_id(),
                 true,
@@ -2406,15 +2426,10 @@ mod set_local_media_settings {
         );
         mock_navigator.stop();
 
-        assert_eq!(&err.name(), "RecoveredException");
-        let err = err.recover_reason().unwrap();
-
-        assert_eq!(err.name(), "CouldNotGetLocalMedia");
-        assert_eq!(
-            err.message(),
-            "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Unknown JS error: disables_on_fail"
-        );
+        let cause: LocalMediaInitException =
+            unchecked_jsval_cast(err.cause().into());
+        assert!(cause.message().contains("disables_on_fail"));
+        assert_eq!(err.rolled_back(), false);
 
         assert!(!peer1.is_send_video_enabled(Some(MediaSourceKind::Device)));
         assert!(peer1.get_send_tracks().is_empty());
@@ -2436,7 +2451,7 @@ mod set_local_media_settings {
         .unwrap();
 
         let mock_navigator = MockNavigator::new();
-        let err = get_constraints_update_exception(
+        let err: MediaSettingsUpdateException = unchecked_jsval_cast(
             JsFuture::from(room_handle.set_local_media_settings(
                 &media_settings_with_device_id(),
                 true,
@@ -2447,12 +2462,14 @@ mod set_local_media_settings {
         );
         mock_navigator.stop();
 
-        assert_eq!(err.name(), "RecoveredException");
-        let recover_reason = err.recover_reason().unwrap();
-        assert_eq!(recover_reason.name(), "CouldNotGetLocalMedia");
+        let cause: LocalMediaInitException =
+            unchecked_jsval_cast(err.cause().into());
+        assert_eq!(err.rolled_back(), true);
+        assert!(cause.message().contains(
+            "Failed to get local tracks: MediaDevices.getUserMedia() failed",
+        ));
 
         assert_eq!(mock_navigator.get_user_media_requests_count(), 2);
-
         assert!(peer1.is_send_video_enabled(Some(MediaSourceKind::Device)));
         assert_eq!(peer1.get_send_tracks().len(), 1);
     }
@@ -2474,7 +2491,7 @@ mod set_local_media_settings {
 
         let mock_navigator = MockNavigator::new();
         mock_navigator.error_get_user_media("disables_on_rollback_fail".into());
-        let err = get_constraints_update_exception(
+        let err: MediaSettingsUpdateException = unchecked_jsval_cast(
             JsFuture::from(room_handle.set_local_media_settings(
                 &media_settings_with_device_id(),
                 true,
@@ -2485,25 +2502,12 @@ mod set_local_media_settings {
         );
         mock_navigator.stop();
 
-        assert_eq!(err.name(), "RecoverFailedException");
-        let recover_reason = err.recover_reason().unwrap();
-        assert_eq!(recover_reason.name(), "CouldNotGetLocalMedia");
-        assert_eq!(
-            recover_reason.message(),
-            "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Unknown JS error: disables_on_rollback_fail"
-        );
-        let recover_fail_reasons =
-            js_sys::Array::from(&err.recover_fail_reasons());
-        assert_eq!(recover_fail_reasons.length(), 1);
-        let recover_fail_reason = err.recover_reason().unwrap();
-        assert_eq!(
-            recover_fail_reason.message(),
-            "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Unknown JS error: disables_on_rollback_fail"
-        );
-        assert_eq!(recover_fail_reason.name(), "CouldNotGetLocalMedia");
-
+        let cause: LocalMediaInitException =
+            unchecked_jsval_cast(err.cause().into());
+        assert_eq!(err.rolled_back(), false);
+        assert!(cause.message().contains(
+            "Failed to get local tracks: MediaDevices.getUserMedia() failed",
+        ));
         assert!(!peer1.is_send_video_enabled(Some(MediaSourceKind::Device)));
         assert!(peer1.get_send_tracks().is_empty());
     }
@@ -2527,7 +2531,7 @@ mod set_local_media_settings {
         mock_navigator
             .error_get_user_media("doesnt_disables_if_not_stop_first".into());
 
-        let err = get_constraints_update_exception(
+        let err: MediaSettingsUpdateException = unchecked_jsval_cast(
             JsFuture::from(room_handle.set_local_media_settings(
                 &media_settings_with_device_id(),
                 false,
@@ -2538,9 +2542,12 @@ mod set_local_media_settings {
         );
         mock_navigator.stop();
 
-        assert_eq!(err.name(), "RecoveredException");
-        let err = err.recover_reason().unwrap();
-        assert_eq!(err.name(), "CouldNotGetLocalMedia");
+        let cause: LocalMediaInitException =
+            unchecked_jsval_cast(err.cause().into());
+        assert_eq!(err.rolled_back(), true);
+        assert!(cause.message().contains(
+            "Failed to get local tracks: MediaDevices.getUserMedia() failed",
+        ));
 
         assert!(peer1.is_send_video_enabled(Some(MediaSourceKind::Device)));
         assert_eq!(peer1.get_send_tracks().len(), 1);